+++ zcncore/transaction_query_mobile.go
@@ -6,156 +6,16 @@ package zcncore
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	stderrors "errors"
-	"math/rand"
 	"net/http"
 	"strconv"
-	"strings"
-	"time"
 
 	thrown "github.com/0chain/errors"
 	"github.com/0chain/gosdk/core/resty"
-	"github.com/0chain/gosdk/core/util"
 )
-
-var (
-	ErrNoAvailableSharders     = errors.New("zcn: no available sharders")
-	ErrNoEnoughSharders        = errors.New("zcn: sharders is not enough")
-	ErrNoEnoughOnlineSharders  = errors.New("zcn: online sharders is not enough")
-	ErrInvalidNumSharder       = errors.New("zcn: number of sharders is invalid")
-	ErrNoOnlineSharders        = errors.New("zcn: no any online sharder")
-	ErrSharderOffline          = errors.New("zcn: sharder is offline")
-	ErrInvalidConsensus        = errors.New("zcn: invalid consensus")
-	ErrTransactionNotFound     = errors.New("zcn: transaction not found")
-	ErrTransactionNotConfirmed = errors.New("zcn: transaction not confirmed")
-)
-
-const (
-	SharderEndpointHealthCheck = "/_health_check"
-)
-
-// queryResultHandle handle query response, return true if it is a consensus-result
-type queryResultHandle func(result QueryResult) bool
-
-func (tq *TransactionQuery) Reset() {
-	tq.selected = make(map[string]bool)
-	tq.offline = make(map[string]bool)
-}
-
-// validate validate data and input
-func (tq *TransactionQuery) validate(num int) error {
-	if tq == nil || tq.max == 0 {
-		return ErrNoAvailableSharders
-	}
-
-	if num < 1 {
-		return ErrInvalidNumSharder
-	}
-
-	if num > tq.max {
-		return ErrNoEnoughSharders
-	}
-
-	if num > (tq.max - len(tq.offline)) {
-		return ErrNoEnoughOnlineSharders
-	}
-
-	return nil
-
-}
-
-// buildUrl build url with host and parts
-func (tq *TransactionQuery) buildUrl(host string, parts ...string) string {
-	var sb strings.Builder
-
-	sb.WriteString(strings.TrimSuffix(host, "/"))
-
-	for _, it := range parts {
-		sb.WriteString(it)
-	}
-
-	return sb.String()
-}
-
-// checkHealth check health
-func (tq *TransactionQuery) checkHealth(ctx context.Context, host string) error {
-
-	_, ok := tq.offline[host]
-	if ok {
-		return ErrSharderOffline
-	}
-
-	// check health
-	r := resty.New()
-	requestUrl := tq.buildUrl(host, SharderEndpointHealthCheck)
-	logging.Info("zcn: check health ", requestUrl)
-	r.DoGet(ctx, requestUrl)
-	r.Then(func(req *http.Request, resp *http.Response, respBody []byte, cf context.CancelFunc, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// 5xx: it is a server error, not client error
-		if resp.StatusCode >= http.StatusInternalServerError {
-			return thrown.Throw(ErrSharderOffline, resp.Status)
-		}
-
-		return nil
-	})
-	errs := r.Wait()
-
-	if len(errs) > 0 {
-		tq.offline[host] = true
-
-		if len(tq.offline) >= tq.max {
-			return ErrNoOnlineSharders
-		}
-	}
-
-	return nil
-}
-
-// randOne random one health sharder
-func (tq *TransactionQuery) randOne(ctx context.Context) (string, error) {
-
-	randGen := rand.New(rand.NewSource(time.Now().UnixNano()))
-	for {
-
-		// reset selected if all sharders were selected
-		if len(tq.selected) >= tq.max {
-			tq.selected = make(map[string]bool)
-		}
-
-		i := randGen.Intn(len(tq.sharders))
-		host := tq.sharders[i]
-
-		_, ok := tq.selected[host]
-
-		// it was selected, try next
-		if ok {
-			continue
-		}
-
-		tq.selected[host] = true
-
-		err := tq.checkHealth(ctx, host)
-
-		if err != nil {
-			if errors.Is(err, ErrNoOnlineSharders) {
-				return "", err
-			}
-
-			// it is offline, try next one
-			continue
-		}
-
-		return host, nil
-	}
-}
 
 // fromAll query transaction from all sharders whatever it is selected or offline in previous queires, and return consensus result
-func (tq *TransactionQuery) fromAll(query string, handle queryResultHandle, timeout RequestTimeout) error {
+func (tq *TransactionQuery) fromAll(query string, handle QueryResultHandle, timeout RequestTimeout) error {
 	if tq == nil || tq.max == 0 {
 		return ErrNoAvailableSharders
 	}
@@ -189,7 +49,6 @@ func (tq *TransactionQuery) fromAll(query string, handle queryResultHandle, time
 
 			if handle != nil {
 				if handle(res) {
-
 					cf()
 				}
 			}
@@ -253,50 +112,6 @@ func (tq *TransactionQuery) fromAny(query string, timeout RequestTimeout) (Query
 
 	return res, nil
 
-}
-
-func (tq *TransactionQuery) getInfo(query string, timeout RequestTimeout) (*QueryResult, error) {
-
-	consensuses := make(map[int]int)
-	var maxConsensus int
-	var consensusesResp QueryResult
-	// {host}{query}
-
-	err := tq.fromAll(query,
-		func(qr QueryResult) bool {
-			//ignore response if it is network error
-			if qr.StatusCode >= 500 {
-				return false
-			}
-
-			consensuses[qr.StatusCode]++
-			if consensuses[qr.StatusCode] >= maxConsensus {
-				maxConsensus = consensuses[qr.StatusCode]
-				consensusesResp = qr
-			}
-
-			return false
-
-		}, timeout)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if maxConsensus == 0 {
-		return nil, stderrors.New("zcn: query not found")
-	}
-
-	rate := float32(maxConsensus*100) / float32(tq.max)
-	if rate < consensusThresh {
-		return nil, ErrInvalidConsensus
-	}
-
-	if consensusesResp.StatusCode != http.StatusOK {
-		return nil, stderrors.New(string(consensusesResp.Content))
-	}
-
-	return &consensusesResp, nil
 }
 
 func (tq *TransactionQuery) getConsensusConfirmation(numSharders int, txnHash string, timeout RequestTimeout) (*blockHeader, map[string]json.RawMessage, *blockHeader, error) {
@@ -424,38 +239,4 @@ func (tq *TransactionQuery) getFastConfirmation(txnHash string, timeout RequestT
 	}
 
 	return nil, nil, nil, thrown.Throw(ErrTransactionNotFound, strconv.Itoa(result.StatusCode))
-}
-
-func GetInfoFromSharders(urlSuffix string, op int, cb GetInfoCallback) {
-
-	tq, err := newTransactionQuery(util.Shuffle(_config.chain.Sharders))
-	if err != nil {
-		cb.OnInfoAvailable(op, StatusError, "", err.Error())
-		return
-	}
-
-	qr, err := tq.getInfo(urlSuffix, nil)
-	if err != nil {
-		cb.OnInfoAvailable(op, StatusError, "", err.Error())
-		return
-	}
-
-	cb.OnInfoAvailable(op, StatusSuccess, string(qr.Content), "")
-}
-
-func GetInfoFromAnySharder(urlSuffix string, op int, cb GetInfoCallback) {
-
-	tq, err := newTransactionQuery(util.Shuffle(_config.chain.Sharders))
-	if err != nil {
-		cb.OnInfoAvailable(op, StatusError, "", err.Error())
-		return
-	}
-
-	qr, err := tq.fromAny(urlSuffix, nil)
-	if err != nil {
-		cb.OnInfoAvailable(op, StatusError, "", err.Error())
-		return
-	}
-
-	cb.OnInfoAvailable(op, StatusSuccess, string(qr.Content), "")
 }
